+++ tests/testthat/test-environmnet.R
@@ -22,8 +22,8 @@ g2 <- function(g2v1){
 }
 test_that("Do %__% and %v__% behave appropriately", {
   expect_equal(6 %>>% g1 %>% sapply(m_OK), c(TRUE, TRUE, TRUE, TRUE))
-  expect_equal(6 %>>% g1 %>% lapply(m_value), list(NULL,6,NULL,5))
+  expect_equal(6 %>>% g1 %>% lapply(m_value, warn=FALSE), list(NULL,6,NULL,5))
 
   expect_equal(6 %>>% g2 %>% sapply(m_OK), c(TRUE, TRUE, TRUE, TRUE))
-  expect_equal(6 %>>% g2 %>% lapply(m_value), list(NULL,NULL,NULL,5))
+  expect_equal(6 %>>% g2 %>% lapply(m_value, warn=FALSE), list(NULL,NULL,NULL,5))
 })